requirements.txt
gcsfs
python-dotenv
google-auth